Move Price to the same package as PriceRange
since PriceRange depends on Price, it makes sense that both of them are
in the same package. refactored the test files accordingly as well.
