+++ src/main/java/seedu/address/logic/commands/buyer/EditBuyerCommand.java
@@ -28,7 +28,7 @@ import seedu.address.model.buyer.Name;
 import seedu.address.model.buyer.Phone;
 import seedu.address.model.buyer.Priority;
 import seedu.address.model.characteristics.Characteristics;
-import seedu.address.model.pricerange.PriceRange;
+import seedu.address.model.price.PriceRange;
 
 /**
  * Edits the details of an existing buyer in the address book.
