+++ src/main/java/seedu/address/logic/parser/buyer/AddBuyerCommandParser.java
@@ -24,7 +24,7 @@ import seedu.address.model.buyer.Name;
 import seedu.address.model.buyer.Phone;
 import seedu.address.model.buyer.Priority;
 import seedu.address.model.characteristics.Characteristics;
-import seedu.address.model.pricerange.PriceRange;
+import seedu.address.model.price.PriceRange;
 
 
 /**
