+++ src/main/java/seedu/address/model/price/PriceRange.java
@@ -1,11 +1,9 @@
-package seedu.address.model.pricerange;
+package seedu.address.model.price;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
 
 import java.text.DecimalFormat;
-
-import seedu.address.model.property.Price;
 
 /**
  * Represents a Price Range for a property that a buyer can accept.
