+++ src/main/java/seedu/address/storage/JsonAdaptedBuyer.java
@@ -13,7 +13,7 @@ import seedu.address.model.buyer.Name;
 import seedu.address.model.buyer.Phone;
 import seedu.address.model.buyer.Priority;
 import seedu.address.model.characteristics.Characteristics;
-import seedu.address.model.pricerange.PriceRange;
+import seedu.address.model.price.PriceRange;
 
 /**
  * Jackson-friendly version of {@link Buyer}.
