+++ src/test/java/seedu/address/logic/parser/property/EditPropertyCommandParserTest.java
@@ -53,7 +53,7 @@ import seedu.address.model.buyer.Name;
 import seedu.address.model.buyer.Phone;
 import seedu.address.model.characteristics.Characteristics;
 import seedu.address.model.property.Description;
-import seedu.address.model.property.Price;
+import seedu.address.model.price.Price;
 import seedu.address.model.property.PropertyName;
 import seedu.address.testutil.EditPropertyDescriptorBuilder;
