+++ src/test/java/seedu/address/model/price/PriceTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.property;
+package seedu.address.model.price;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
