+++ src/test/java/seedu/address/testutil/EditBuyerDescriptorBuilder.java
@@ -9,7 +9,7 @@ import seedu.address.model.buyer.Name;
 import seedu.address.model.buyer.Phone;
 import seedu.address.model.buyer.Priority;
 import seedu.address.model.characteristics.Characteristics;
-import seedu.address.model.pricerange.PriceRange;
+import seedu.address.model.price.PriceRange;
 
 /**
  * A utility class to help with building EditBuyerDescriptor objects.
