+++ src/test/java/seedu/address/testutil/PropertyBuilder.java
@@ -8,7 +8,7 @@ import seedu.address.model.buyer.Phone;
 import seedu.address.model.characteristics.Characteristics;
 import seedu.address.model.property.Description;
 import seedu.address.model.property.Owner;
-import seedu.address.model.property.Price;
+import seedu.address.model.price.Price;
 import seedu.address.model.property.Property;
 import seedu.address.model.property.PropertyName;
